pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "whisper-wrapper"
version = "1.0.0"
description = "Lightweight tray app for voice-to-text on Linux using faster-whisper."
readme = "README.md"
requires-python = ">=3.10"
license = {file = "LICENSE"}
authors = [{name = "Whisper Wrapper Contributors"}]
keywords = ["whisper", "speech-to-text", "transcription", "linux", "tray"]
classifiers = [
  "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Operating System :: POSIX :: Linux",
]
dependencies = [
  "PyQt6>=6.10.1,<7.0",
  "faster-whisper>=1.2.1,<2.0",
  "sounddevice>=0.5.3,<0.6",
  "numpy>=2.3.5,<3.0",
  "pynput>=1.8.1,<2.0",
  "pyperclip>=1.11.0,<2.0",
  "webrtcvad>=2.0.10",
  "nvidia-cudnn-cu12>=9.0",
  "nvidia-cublas-cu12>=12.0",
]

[tool.setuptools.packages.find]
include = ["src*"]

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.ruff]
line-length = 88
target-version = "py310"
exclude = ["venv", ".venv", "build", "dist", ".idea", "__pycache__"]

[tool.ruff.lint]
select = ["E", "F"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_unused_ignores = true
show_error_codes = true
pretty = true
ignore_missing_imports = true
files = ["src", "trigger.py"]
exclude = "(venv|\\.venv|build|dist|\\.idea|__pycache__)"

[tool.pytest.ini_options]
testpaths = ["tests"]
